refactor(FormatConstants): separate string values for output format
