+++ src/main/java/vendingmachine/constants/FormatConstants.java
@@ -1,0 +1,9 @@
+package vendingmachine.constants;
+
+public class FormatConstants {
+
+    public static String BLANK_LINE = "\n";
+    public static String WON = "원";
+    public static String COIN_PRINT_SEPARATOR = " - ";
+    public static String COIN_AMOUNT_DIGIT = "개";
+}
