+++ src/main/java/vendingmachine/constants/InputMessages.java
@@ -1,9 +1,11 @@
 package vendingmachine.constants;
 
+import static vendingmachine.constants.FormatConstants.BLANK_LINE;
+
 public class InputMessages {
 
     public static String REQUEST_VENDING_MACHINE_INITIAL_MONEY = "자판기가 보유하고 있는 금액을 입력해 주세요.";
     public static String REQUEST_MENU_INPUT = "상품명과 가격, 수량을 입력해 주세요.";
-    public static String REQUEST_CUSTOMER_MONEY_INPUT = SystemConstants.BLANK_LINE + "투입 금액을 입력해 주세요.";
+    public static String REQUEST_CUSTOMER_MONEY_INPUT = BLANK_LINE + "투입 금액을 입력해 주세요.";
     public static String REQUEST_PURCHASE_INPUT = "구매할 상품명을 입력해 주세요.";
 }
